Allow usage as a library
