+++ README.md
@@ -112,7 +112,7 @@ bun install
 To run:
 
 ```bash
-bun run index.ts
+bun run cli.ts
 ```
 
 ## Contributions
