+++ lib/cli-helpers.ts
